Update parental-leave e2e tests
